+++ src/cycle.js
@@ -67,7 +67,7 @@ var Cycle = {
    *
    *     createReactClass('displayName', definitionFn, {
    *       rootTagName: 'div',
-   *       mixins: [PureRenderMixin],
+   *       mixins: [],
    *       propTypes: null,
    *       disableHotLoader: false,
    *       bindThis: false